src/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.1
h11==0.14.0
idna==3.4
protobuf==4.22.3
psycopg2-binary==2.9.6
pydantic==1.10.2
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.15.0
